Word Embeddings Builder
Saves word embedding model to a file. This feature is commented out
currently because I have no need to save a model for the uncleaned train
data (no whitespace removal, etc). To get it working with the cleaned
data, just change the read_csv file (line 7) and uncomment the
model.save line (line 25).